python-rag-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.3.1
openai==1.10.0
httpx==0.26.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
pandas==2.1.4
